requirements.txt
datasets==2.8.0
evaluate==0.4.0
Flask==2.2.2
Flask_Cors==3.0.10
graphviz==0.20.1
numpy==1.24.0
python-dotenv==1.0.0
requests==2.28.1
sentence_transformers==2.2.2
spacy==3.4.4
supabase==1.0.1
transformers==4.25.1
